# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An example of gin",
        "title": "Golang Gin API",
        "termsOfService": "https://github.com/EDDYCJY/go-gin-example",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/EDDYCJY/go-gin-example/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/assetSearch": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get asset details of an agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id",
                        "name": "taskid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "agent id",
                        "name": "agentid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/getAgentList": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get agent list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/getProgress": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get progress of a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id",
                        "name": "taskid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "agent id",
                        "name": "agentid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/getTaskAttr": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get task attr",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id",
                        "name": "taskid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/getTaskList": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get task list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/startScan": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Start scan task to the pointed agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "agent id",
                        "name": "agentid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ip segment",
                        "name": "ips",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}